Creating a BuildLauncher class used on the home and the builds screens
